Move graphics tools to sstudio.graphics.
