+++ smarts/sstudio/graphics/heightfield.py
