+++ smarts/sstudio/graphics/mesh2bitmap.py
